refactor(aztec-node): extract node creation into a createAztecNodeService factory
Move the ~580-line `AztecNodeService.createAndSync` static method (and its
`checkConfigMatchesRollup` helper) out of `server.ts` into a new `factory.ts` as a
free function `createAztecNodeService`, mirroring the `createProverNode` /
`createValidatorClient` factories in other subsystems.

Convert the `AztecNodeService` constructor from ~28 positional parameters to a single
`AztecNodeServiceDeps` object so call sites name each dependency. This surfaced a latent
bug in the TXE state machine, which passed `VERSION` and `CHAIN_ID` into the `l1ChainId`
and `version` slots in the wrong order (both happened to be `1`); the named object fixes it.

Update all `createAndSync` call sites (node bin, local-network, e2e fixtures) and the
direct constructor call in the TXE state machine. `server.ts` drops ~590 lines and a large
set of now-unused imports.
