+++ yarn-project/aztec-node/src/aztec-node/server.ts
@@ -1,15 +1,11 @@
-import { Archiver, createArchiver } from '@aztec/archiver';
+import { Archiver } from '@aztec/archiver';
 import { BBCircuitVerifier, BatchChonkVerifier, QueuedIVCVerifier } from '@aztec/bb-prover';
 import { TestCircuitVerifier } from '@aztec/bb-prover/test';
-import { type BlobClientInterface, createBlobClientWithFileStores } from '@aztec/blob-client/client';
-import { Blob } from '@aztec/blob-lib';
+import type { BlobClientInterface } from '@aztec/blob-client/client';
 import { ARCHIVE_HEIGHT, type L1_TO_L2_MSG_TREE_HEIGHT, type NOTE_HASH_TREE_HEIGHT } from '@aztec/constants';
-import { EpochCache, type EpochCacheInterface } from '@aztec/epoch-cache';
-import { createEthereumChain } from '@aztec/ethereum/chain';
-import { getPublicClient, makeL1HttpTransport } from '@aztec/ethereum/client';
-import { RegistryContract, RollupContract } from '@aztec/ethereum/contracts';
+import type { EpochCacheInterface } from '@aztec/epoch-cache';
+import { RollupContract } from '@aztec/ethereum/contracts';
 import { type L1ContractAddresses, pickL1ContractAddresses } from '@aztec/ethereum/l1-contract-addresses';
-import type { L1TxUtils } from '@aztec/ethereum/l1-tx-utils';
 import {
   BlockNumber,
   CheckpointNumber,
@@ -24,38 +20,16 @@ import { BadRequestError } from '@aztec/foundation/json-rpc';
 import { type Logger, createLogger } from '@aztec/foundation/log';
 import { retryUntil } from '@aztec/foundation/retry';
 import { count } from '@aztec/foundation/string';
-import { DateProvider, Timer } from '@aztec/foundation/timer';
+import { Timer } from '@aztec/foundation/timer';
 import { MembershipWitness, SiblingPath } from '@aztec/foundation/trees';
-import { type KeyStore, KeystoreManager, loadKeystores, mergeKeystores } from '@aztec/node-keystore';
-import { trySnapshotSync, uploadSnapshot } from '@aztec/node-lib/actions';
-import { createForwarderL1TxUtilsFromSigners, createL1TxUtilsFromSigners } from '@aztec/node-lib/factories';
-import {
-  type P2P,
-  type P2PClientDeps,
-  createP2PClient,
-  createTxValidatorForAcceptingTxsOverRPC,
-  getDefaultAllowedSetupFunctions,
-} from '@aztec/p2p';
+import { KeystoreManager, loadKeystores, mergeKeystores } from '@aztec/node-keystore';
+import { uploadSnapshot } from '@aztec/node-lib/actions';
+import { type P2P, createTxValidatorForAcceptingTxsOverRPC, getDefaultAllowedSetupFunctions } from '@aztec/p2p';
 import { ProtocolContractAddress } from '@aztec/protocol-contracts';
-import { type ProverNode, type ProverNodeDeps, createProverNode } from '@aztec/prover-node';
-import { createKeyStoreForProver } from '@aztec/prover-node/config';
-import {
-  FeeProviderImpl,
-  GlobalVariableBuilder,
-  SequencerClient,
-  type SequencerPublisher,
-} from '@aztec/sequencer-client';
-import { AutomineSequencer, createAutomineSequencer } from '@aztec/sequencer-client/automine';
-import {
-  AttestationsBlockWatcher,
-  AttestedInvalidProposalWatcher,
-  BroadcastedInvalidCheckpointProposalWatcher,
-  CheckpointEquivocationWatcher,
-  DataWithholdingWatcher,
-  type SlasherClientInterface,
-  type Watcher,
-  createSlasher,
-} from '@aztec/slasher';
+import type { ProverNode } from '@aztec/prover-node';
+import { SequencerClient } from '@aztec/sequencer-client';
+import { AutomineSequencer } from '@aztec/sequencer-client/automine';
+import type { SlasherClientInterface } from '@aztec/slasher';
 import { STANDARD_MULTI_CALL_ENTRYPOINT_ADDRESS } from '@aztec/standard-contracts/multi-call-entrypoint';
 import { AztecAddress } from '@aztec/stdlib/aztec-address';
 import {
@@ -72,12 +46,7 @@ import {
   type NormalizedBlockParameter,
   inspectBlockParameter,
 } from '@aztec/stdlib/block';
-import {
-  type CheckpointData,
-  CheckpointReexecutionTracker,
-  L1PublishedData,
-  type PublishedCheckpoint,
-} from '@aztec/stdlib/checkpoint';
+import { type CheckpointData, L1PublishedData, type PublishedCheckpoint } from '@aztec/stdlib/checkpoint';
 import type {
   ContractClassPublic,
   ContractDataSource,
@@ -115,7 +84,7 @@ import {
   tryStop,
 } from '@aztec/stdlib/interfaces/server';
 import type { DebugLogStore, LogResult, PrivateLogsQuery, PublicLogsQuery } from '@aztec/stdlib/logs';
-import { InMemoryDebugLogStore, NullDebugLogStore } from '@aztec/stdlib/logs';
+import { NullDebugLogStore } from '@aztec/stdlib/logs';
 import { InboxLeaf, type L1ToL2MessageSource, type L2ToL1MembershipWitness } from '@aztec/stdlib/messaging';
 import type { CheckpointAttestation } from '@aztec/stdlib/p2p';
 import type { Offense } from '@aztec/stdlib/slashing';
@@ -138,9 +107,7 @@ import {
   TxStatus,
   type TxValidationResult,
 } from '@aztec/stdlib/tx';
-import { getPackageVersion } from '@aztec/stdlib/update-checker';
 import type { SingleValidatorStats, ValidatorsStats } from '@aztec/stdlib/validators';
-import type { GenesisData } from '@aztec/stdlib/world-state';
 import {
   Attributes,
   type TelemetryClient,
@@ -149,21 +116,8 @@ import {
   getTelemetryClient,
   trackSpan,
 } from '@aztec/telemetry-client';
-import {
-  FullNodeCheckpointsBuilder as CheckpointsBuilder,
-  FullNodeCheckpointsBuilder,
-  NodeKeystoreAdapter,
-  ProposalHandler,
-  ValidatorClient,
-  createProposalHandler,
-  createValidatorClient,
-} from '@aztec/validator-client';
-import type { SlashingProtectionDatabase } from '@aztec/validator-ha-signer/types';
-import { createWorldState, createWorldStateSynchronizer } from '@aztec/world-state';
+import { NodeKeystoreAdapter, ValidatorClient } from '@aztec/validator-client';
 
-import { createPublicClient } from 'viem';
-
-import { createSentinel } from '../sentinel/factory.js';
 import { Sentinel } from '../sentinel/sentinel.js';
 import {
   blockResponseFromBlockData,
@@ -172,9 +126,45 @@ import {
   checkpointResponseFromPublishedCheckpoint,
   projectProposedToCheckpointResponse,
 } from './block_response_helpers.js';
-import { type AztecNodeConfig, createKeyStoreForValidator } from './config.js';
+import type { AztecNodeConfig } from './config.js';
 import { NodeMetrics } from './node_metrics.js';
 import { NodePublicCallsSimulator } from './node_public_calls_simulator.js';
+
+/**
+ * Fully-constructed collaborators and settings an {@link AztecNodeService} owns. Built by `createAztecNodeService`
+ * (see `factory.ts`); passed as a single object so call sites name each dependency instead of relying on
+ * positional order.
+ */
+export interface AztecNodeServiceDeps {
+  config: AztecNodeConfig;
+  p2pClient: P2P;
+  blockSource: L2BlockSource & Partial<Service>;
+  logsSource: L2LogsSource;
+  contractDataSource: ContractDataSource;
+  l1ToL2MessageSource: L1ToL2MessageSource;
+  worldStateSynchronizer: WorldStateSynchronizer;
+  sequencer: SequencerClient | undefined;
+  proverNode: ProverNode | undefined;
+  slasherClient: SlasherClientInterface | undefined;
+  validatorsSentinel: Sentinel | undefined;
+  stopStartedWatchers: () => Promise<void>;
+  l1ChainId: number;
+  version: number;
+  globalVariableBuilder: GlobalVariableBuilderInterface;
+  rollupContract: RollupContract | undefined;
+  feeProvider: FeeProvider;
+  epochCache: EpochCacheInterface;
+  packageVersion: string;
+  peerProofVerifier: ClientProtocolCircuitVerifier;
+  rpcProofVerifier: ClientProtocolCircuitVerifier;
+  telemetry?: TelemetryClient;
+  log?: Logger;
+  blobClient?: BlobClientInterface;
+  validatorClient?: ValidatorClient;
+  keyStoreManager?: KeystoreManager;
+  debugLogStore?: DebugLogStore;
+  automineSequencer?: AutomineSequencer;
+}
 
 /**
  * The aztec node.
@@ -189,38 +179,67 @@ export class AztecNodeService implements AztecNode, AztecNodeAdmin, AztecNodeDeb
 
   public readonly tracer: Tracer;
 
-  constructor(
-    protected config: AztecNodeConfig,
-    protected readonly p2pClient: P2P,
-    protected readonly blockSource: L2BlockSource & Partial<Service>,
-    protected readonly logsSource: L2LogsSource,
-    protected readonly contractDataSource: ContractDataSource,
-    protected readonly l1ToL2MessageSource: L1ToL2MessageSource,
-    protected readonly worldStateSynchronizer: WorldStateSynchronizer,
-    protected readonly sequencer: SequencerClient | undefined,
-    protected readonly proverNode: ProverNode | undefined,
-    protected readonly slasherClient: SlasherClientInterface | undefined,
-    protected readonly validatorsSentinel: Sentinel | undefined,
-    private readonly stopStartedWatchers: () => Promise<void>,
-    protected readonly l1ChainId: number,
-    protected readonly version: number,
-    protected readonly globalVariableBuilder: GlobalVariableBuilderInterface,
-    protected readonly rollupContract: RollupContract | undefined,
-    protected readonly feeProvider: FeeProvider,
-    protected readonly epochCache: EpochCacheInterface,
-    protected readonly packageVersion: string,
-    private peerProofVerifier: ClientProtocolCircuitVerifier,
-    private rpcProofVerifier: ClientProtocolCircuitVerifier,
-    private telemetry: TelemetryClient = getTelemetryClient(),
-    private log = createLogger('node'),
-    private blobClient?: BlobClientInterface,
-    private validatorClient?: ValidatorClient,
-    private keyStoreManager?: KeystoreManager,
-    private debugLogStore: DebugLogStore = new NullDebugLogStore(),
-    private readonly automineSequencer?: AutomineSequencer,
-  ) {
-    this.metrics = new NodeMetrics(telemetry, 'AztecNodeService');
-    this.tracer = telemetry.getTracer('AztecNodeService');
+  protected config: AztecNodeConfig;
+  protected readonly p2pClient: P2P;
+  protected readonly blockSource: L2BlockSource & Partial<Service>;
+  protected readonly logsSource: L2LogsSource;
+  protected readonly contractDataSource: ContractDataSource;
+  protected readonly l1ToL2MessageSource: L1ToL2MessageSource;
+  protected readonly worldStateSynchronizer: WorldStateSynchronizer;
+  protected readonly sequencer: SequencerClient | undefined;
+  protected readonly proverNode: ProverNode | undefined;
+  protected readonly slasherClient: SlasherClientInterface | undefined;
+  protected readonly validatorsSentinel: Sentinel | undefined;
+  private readonly stopStartedWatchers: () => Promise<void>;
+  protected readonly l1ChainId: number;
+  protected readonly version: number;
+  protected readonly globalVariableBuilder: GlobalVariableBuilderInterface;
+  protected readonly rollupContract: RollupContract | undefined;
+  protected readonly feeProvider: FeeProvider;
+  protected readonly epochCache: EpochCacheInterface;
+  protected readonly packageVersion: string;
+  private peerProofVerifier: ClientProtocolCircuitVerifier;
+  private rpcProofVerifier: ClientProtocolCircuitVerifier;
+  private telemetry: TelemetryClient;
+  private log: Logger;
+  private blobClient?: BlobClientInterface;
+  private validatorClient?: ValidatorClient;
+  private keyStoreManager?: KeystoreManager;
+  private debugLogStore: DebugLogStore;
+  private readonly automineSequencer?: AutomineSequencer;
+
+  constructor(deps: AztecNodeServiceDeps) {
+    this.config = deps.config;
+    this.p2pClient = deps.p2pClient;
+    this.blockSource = deps.blockSource;
+    this.logsSource = deps.logsSource;
+    this.contractDataSource = deps.contractDataSource;
+    this.l1ToL2MessageSource = deps.l1ToL2MessageSource;
+    this.worldStateSynchronizer = deps.worldStateSynchronizer;
+    this.sequencer = deps.sequencer;
+    this.proverNode = deps.proverNode;
+    this.slasherClient = deps.slasherClient;
+    this.validatorsSentinel = deps.validatorsSentinel;
+    this.stopStartedWatchers = deps.stopStartedWatchers;
+    this.l1ChainId = deps.l1ChainId;
+    this.version = deps.version;
+    this.globalVariableBuilder = deps.globalVariableBuilder;
+    this.rollupContract = deps.rollupContract;
+    this.feeProvider = deps.feeProvider;
+    this.epochCache = deps.epochCache;
+    this.packageVersion = deps.packageVersion;
+    this.peerProofVerifier = deps.peerProofVerifier;
+    this.rpcProofVerifier = deps.rpcProofVerifier;
+    this.telemetry = deps.telemetry ?? getTelemetryClient();
+    this.log = deps.log ?? createLogger('node');
+    this.blobClient = deps.blobClient;
+    this.validatorClient = deps.validatorClient;
+    this.keyStoreManager = deps.keyStoreManager;
+    this.debugLogStore = deps.debugLogStore ?? new NullDebugLogStore();
+    this.automineSequencer = deps.automineSequencer;
+
+    this.metrics = new NodeMetrics(this.telemetry, 'AztecNodeService');
+    this.tracer = this.telemetry.getTracer('AztecNodeService');
 
     // The node never represents a proposer's payout addresses, so the simulator zeroes coinbase and
     // fee recipient. The signature context only needs chain id + rollup address (see signature_utils).
@@ -239,12 +258,12 @@ export class AztecNodeService implements AztecNode, AztecNodeAdmin, AztecNodeDeb
     });
 
     this.log.info(`Aztec Node version: ${this.packageVersion}`);
-    this.log.info(`Aztec Node started on chain 0x${l1ChainId.toString(16)}`, pickL1ContractAddresses(config));
+    this.log.info(`Aztec Node started on chain 0x${this.l1ChainId.toString(16)}`, pickL1ContractAddresses(this.config));
 
-    // A defensive check that protects us against introducing a bug in the complex `createAndSync` function. We must
+    // A defensive check that protects us against introducing a bug in the complex node creation flow. We must
     // never have debugLogStore enabled when not in test mode because then we would be accumulating debug logs in
     // memory which could be a DoS vector on the sequencer (since no fees are paid for debug logs).
-    if (debugLogStore.isEnabled && config.realProofs) {
+    if (this.debugLogStore.isEnabled && this.config.realProofs) {
       throw new Error('debugLogStore should never be enabled when realProofs are set');
     }
   }
@@ -502,598 +521,6 @@ export class AztecNodeService implements AztecNode, AztecNodeAdmin, AztecNodeDeb
     }
     const datas = await this.blockSource.getCheckpointsData({ from, limit });
     return datas.map(d => checkpointResponseFromCheckpointData(d, options)) as CheckpointResponse<Opts>[];
-  }
-
-  /**
-   * initializes the Aztec Node, wait for component to sync.
-   * @param config - The configuration to be used by the aztec node.
-   * @returns - A fully synced Aztec Node for use in development/testing.
-   */
-  public static async createAndSync(
-    inputConfig: AztecNodeConfig,
-    deps: {
-      telemetry?: TelemetryClient;
-      logger?: Logger;
-      publisher?: SequencerPublisher;
-      dateProvider?: DateProvider;
-      p2pClientDeps?: P2PClientDeps;
-      proverNodeDeps?: Partial<ProverNodeDeps>;
-      slashingProtectionDb?: SlashingProtectionDatabase;
-    } = {},
-    options: {
-      genesis?: GenesisData;
-      dontStartSequencer?: boolean;
-      dontStartProverNode?: boolean;
-    } = {},
-  ): Promise<AztecNodeService> {
-    const config = { ...inputConfig }; // Copy the config so we dont mutate the input object
-    const log = deps.logger ?? createLogger('node');
-
-    // Initialise the bb.js sync WASM singleton here, before any subsystem runs.
-    const { BarretenbergSync } = await import('@aztec/bb.js');
-    await BarretenbergSync.initSingleton();
-
-    const packageVersion = getPackageVersion();
-    const telemetry = deps.telemetry ?? getTelemetryClient();
-    const dateProvider = deps.dateProvider ?? new DateProvider();
-    const ethereumChain = createEthereumChain(config.l1RpcUrls, config.l1ChainId);
-
-    // Build a key store from file if given or from environment otherwise.
-    // We keep the raw KeyStore available so we can merge with prover keys if enableProverNode is set.
-    let keyStoreManager: KeystoreManager | undefined;
-    const keyStoreProvided = config.keyStoreDirectory !== undefined && config.keyStoreDirectory.length > 0;
-    if (keyStoreProvided) {
-      const keyStores = loadKeystores(config.keyStoreDirectory!);
-      keyStoreManager = new KeystoreManager(mergeKeystores(keyStores));
-    } else {
-      const rawKeyStores: KeyStore[] = [];
-      const validatorKeyStore = createKeyStoreForValidator(config);
-      if (validatorKeyStore) {
-        rawKeyStores.push(validatorKeyStore);
-      }
-      if (config.enableProverNode) {
-        const proverKeyStore = createKeyStoreForProver(config);
-        if (proverKeyStore) {
-          rawKeyStores.push(proverKeyStore);
-        }
-      }
-      if (rawKeyStores.length > 0) {
-        keyStoreManager = new KeystoreManager(
-          rawKeyStores.length === 1 ? rawKeyStores[0] : mergeKeystores(rawKeyStores),
-        );
-      }
-    }
-
-    await keyStoreManager?.validateSigners();
-
-    // If we are a validator, verify our configuration before doing too much more.
-    if (!config.disableValidator) {
-      if (keyStoreManager === undefined) {
-        throw new Error('Failed to create key store, a requirement for running a validator');
-      }
-      if (!keyStoreProvided && process.env.NODE_ENV !== 'test') {
-        log.warn("Keystore created from env: it's recommended to use a file-based key store for production");
-      }
-      ValidatorClient.validateKeyStoreConfiguration(keyStoreManager, log);
-    }
-
-    // validate that the actual chain id matches that specified in configuration
-    if (config.l1ChainId !== ethereumChain.chainInfo.id) {
-      throw new Error(
-        `RPC URL configured for chain id ${ethereumChain.chainInfo.id} but expected id ${config.l1ChainId}`,
-      );
-    }
-
-    const publicClient = createPublicClient({
-      chain: ethereumChain.chainInfo,
-      transport: makeL1HttpTransport(config.l1RpcUrls, { timeout: config.l1HttpTimeoutMS }),
-      pollingInterval: config.viemPollingIntervalMS,
-    });
-
-    const l1ContractsAddresses = await RegistryContract.collectAddresses(
-      publicClient,
-      config.registryAddress,
-      config.rollupVersion ?? 'canonical',
-    );
-
-    Object.assign(config, l1ContractsAddresses);
-
-    const rollupContract = new RollupContract(publicClient, config.rollupAddress.toString());
-    const [l1GenesisTime, slotDuration, epochDuration, rollupVersionFromRollup, rollupManaLimit] = await Promise.all([
-      rollupContract.getL1GenesisTime(),
-      rollupContract.getSlotDuration(),
-      rollupContract.getEpochDuration(),
-      rollupContract.getVersion(),
-      rollupContract.getManaLimit().then(Number),
-    ] as const);
-
-    config.rollupVersion ??= Number(rollupVersionFromRollup);
-
-    if (config.rollupVersion !== Number(rollupVersionFromRollup)) {
-      log.warn(
-        `Registry looked up and returned a rollup with version (${config.rollupVersion}), but this does not match with version detected from the rollup directly: (${rollupVersionFromRollup}).`,
-      );
-    }
-
-    const blobClient = await createBlobClientWithFileStores(config, log.createChild('blob-client'));
-
-    // attempt snapshot sync if possible
-    await trySnapshotSync(config, log);
-
-    const epochCache = await EpochCache.create(config.rollupAddress, config, { dateProvider });
-
-    // Track started resources so we can clean up on partial failure during node creation.
-    const started: { stop?(): Promise<void> | void }[] = [];
-    try {
-      config.skipOrphanProposedBlockPruning ||= !!config.useAutomineSequencer;
-
-      AztecNodeService.checkConfigMatchesRollup(config, {
-        slotDuration: Number(slotDuration),
-        epochDuration: Number(epochDuration),
-      });
-
-      // Create world-state first so we can retrieve the initial header before constructing the archiver.
-      const nativeWs = await createWorldState(config, options.genesis);
-      const initialHeader = nativeWs.getInitialHeader();
-      const initialBlockHash = await initialHeader.hash();
-      const archiver = await createArchiver(
-        config,
-        { blobClient, epochCache, telemetry, dateProvider },
-        { blockUntilSync: !config.skipArchiverInitialSync },
-        initialHeader,
-        initialBlockHash,
-      );
-      started.push(archiver);
-
-      // The synchronizer takes ownership of the native world-state from here
-      const worldStateSynchronizer = await createWorldStateSynchronizer(config, archiver, nativeWs, telemetry);
-      started.push(worldStateSynchronizer);
-      const useRealVerifiers = config.realProofs || config.debugForceTxProofVerification;
-      let peerProofVerifier: ClientProtocolCircuitVerifier;
-      let rpcProofVerifier: ClientProtocolCircuitVerifier;
-      if (useRealVerifiers) {
-        peerProofVerifier = await BatchChonkVerifier.new(config, config.bbChonkVerifyMaxBatch, 'peer');
-        const rpcVerifier = await BBCircuitVerifier.new(config);
-        rpcProofVerifier = new QueuedIVCVerifier(rpcVerifier, config.numConcurrentIVCVerifiers);
-      } else {
-        peerProofVerifier = new TestCircuitVerifier(config.proverTestVerificationDelayMs);
-        rpcProofVerifier = new TestCircuitVerifier(config.proverTestVerificationDelayMs);
-      }
-      started.push(peerProofVerifier, rpcProofVerifier);
-
-      let debugLogStore: DebugLogStore;
-      if (!config.realProofs) {
-        log.warn(`Aztec node is accepting fake proofs`);
-
-        debugLogStore = new InMemoryDebugLogStore();
-        log.info(
-          'Aztec node started in test mode (realProofs set to false) hence debug logs from public functions will be collected and served',
-        );
-      } else {
-        debugLogStore = new NullDebugLogStore();
-      }
-
-      const globalVariableBuilderConfig = {
-        rollupAddress: config.rollupAddress,
-        ethereumSlotDuration: config.ethereumSlotDuration,
-        rollupVersion: BigInt(config.rollupVersion),
-        l1GenesisTime,
-        slotDuration: Number(slotDuration),
-      };
-
-      const globalVariableBuilder = new GlobalVariableBuilder(publicClient, globalVariableBuilderConfig);
-      const feeProvider = new FeeProviderImpl(dateProvider, publicClient, globalVariableBuilderConfig);
-
-      const collectOffenses = !config.disableValidator || config.enableOffenseCollection;
-
-      // create the tx pool and the p2p client, which will need the l2 block source
-      const p2pClient = await createP2PClient(
-        config,
-        archiver,
-        peerProofVerifier,
-        worldStateSynchronizer,
-        epochCache,
-        feeProvider,
-        packageVersion,
-        dateProvider,
-        telemetry,
-        deps.p2pClientDeps,
-        initialBlockHash,
-      );
-      started.push(p2pClient);
-      archiver.setCheckpointProposalPresence(p2pClient);
-
-      // We'll accumulate sentinel watchers here
-      const watchers: Watcher[] = [];
-
-      // Create FullNodeCheckpointsBuilder for block proposal handling and tx validation.
-      // Override maxTxsPerCheckpoint with the validator-specific limit if set.
-      const validatorCheckpointsBuilder = new FullNodeCheckpointsBuilder(
-        {
-          ...config,
-          l1GenesisTime,
-          slotDuration: Number(slotDuration),
-          rollupManaLimit,
-          maxTxsPerCheckpoint: config.validateMaxTxsPerCheckpoint,
-        },
-        worldStateSynchronizer,
-        archiver,
-        dateProvider,
-        telemetry,
-      );
-
-      let validatorClient: ValidatorClient | undefined;
-
-      // The proposal handler (validator-owned or standalone) tracks invalid-proposal/equivocation slots and
-      // feeds the attested-invalid-proposal watcher, so the watcher works on non-validator nodes too.
-      let proposalHandler: ProposalHandler | undefined;
-
-      // Tracks successful checkpoint re-execution by a checkpoint proposal handler.
-      const reexecutionTracker = new CheckpointReexecutionTracker();
-
-      if (!config.disableValidator) {
-        // Create validator client if required
-        validatorClient = await createValidatorClient(config, {
-          checkpointsBuilder: validatorCheckpointsBuilder,
-          worldState: worldStateSynchronizer,
-          p2pClient,
-          telemetry,
-          dateProvider,
-          epochCache,
-          blockSource: archiver,
-          l1ToL2MessageSource: archiver,
-          keyStoreManager,
-          blobClient,
-          reexecutionTracker,
-          slashingProtectionDb: deps.slashingProtectionDb,
-        });
-
-        // If we have a validator client, register it as a source of offenses for the slasher,
-        // and have it register callbacks on the p2p client *before* we start it, otherwise messages
-        // like attestations or auths will fail.
-        if (validatorClient) {
-          watchers.push(validatorClient);
-
-          const vc = validatorClient;
-          const getValidatorAddresses = () => vc.getValidatorAddresses().map(a => a.toString());
-          proposalHandler = validatorClient.getProposalHandler();
-          proposalHandler.register(p2pClient, true, archiver, getValidatorAddresses);
-
-          if (!options.dontStartSequencer) {
-            await validatorClient.registerHandlers();
-          }
-        }
-      }
-
-      // If there's no validator client, create a ProposalHandler to handle block and checkpoint proposals
-      // for monitoring or reexecution. Reexecution (default) allows us to follow the pending chain,
-      // while non-reexecution is used for validating the proposals and collecting their txs.
-      // Checkpoint proposals rebuild blobs if the blob client can upload blobs.
-      if (!validatorClient) {
-        const reexecute = !!config.alwaysReexecuteBlockProposals;
-        log.info(`Setting up proposal handler` + (reexecute ? ' with reexecution of proposals' : ''));
-        proposalHandler = createProposalHandler(config, {
-          checkpointsBuilder: validatorCheckpointsBuilder,
-          worldState: worldStateSynchronizer,
-          epochCache,
-          blockSource: archiver,
-          l1ToL2MessageSource: archiver,
-          p2pClient,
-          blobClient,
-          dateProvider,
-          telemetry,
-          reexecutionTracker,
-        });
-        proposalHandler.register(p2pClient, reexecute, archiver);
-      }
-
-      // Start world state and wait for it to sync to the archiver.
-      await worldStateSynchronizer.start();
-
-      // Start p2p. Note that it depends on world state to be running.
-      await p2pClient.start();
-
-      let dataWithholdingWatcher: DataWithholdingWatcher | undefined;
-      let attestationsBlockWatcher: AttestationsBlockWatcher | undefined;
-      let attestedInvalidProposalWatcher: AttestedInvalidProposalWatcher | undefined;
-      let broadcastedInvalidCheckpointProposalWatcher: BroadcastedInvalidCheckpointProposalWatcher | undefined;
-      let checkpointEquivocationWatcher: CheckpointEquivocationWatcher | undefined;
-
-      const validatorsSentinel = await createSentinel(epochCache, archiver, p2pClient, reexecutionTracker, config);
-      if (validatorsSentinel) {
-        watchers.push(validatorsSentinel);
-      }
-
-      if (collectOffenses) {
-        dataWithholdingWatcher = new DataWithholdingWatcher(
-          epochCache,
-          archiver,
-          p2pClient.getTxProvider(),
-          p2pClient,
-          reexecutionTracker,
-          { chainId: config.l1ChainId, rollupAddress: config.rollupAddress },
-          config,
-        );
-        watchers.push(dataWithholdingWatcher);
-
-        broadcastedInvalidCheckpointProposalWatcher = new BroadcastedInvalidCheckpointProposalWatcher(
-          p2pClient,
-          archiver,
-          epochCache,
-          config,
-        );
-        watchers.push(broadcastedInvalidCheckpointProposalWatcher);
-
-        // The proposal handler (validator-owned or standalone) is the source of invalid-proposal/equivocation
-        // slots, so the watcher runs on non-validator offense collectors too.
-        if (proposalHandler) {
-          attestedInvalidProposalWatcher = new AttestedInvalidProposalWatcher(
-            p2pClient,
-            proposalHandler,
-            archiver,
-            epochCache,
-            config,
-            { log: log.createChild('attested-invalid-proposal-watcher') },
-          );
-          watchers.push(attestedInvalidProposalWatcher);
-        }
-
-        checkpointEquivocationWatcher = new CheckpointEquivocationWatcher(archiver, epochCache, config);
-        watchers.push(checkpointEquivocationWatcher);
-
-        attestationsBlockWatcher = new AttestationsBlockWatcher(archiver, epochCache, config, log.getBindings());
-        watchers.push(attestationsBlockWatcher);
-      }
-
-      const watchersToStart = compactArray([
-        validatorsSentinel,
-        dataWithholdingWatcher,
-        attestationsBlockWatcher,
-        broadcastedInvalidCheckpointProposalWatcher,
-        attestedInvalidProposalWatcher,
-        checkpointEquivocationWatcher,
-      ]);
-      const startedWatchers: Watcher[] = [];
-      const stopStartedWatchers = async () => {
-        for (const watcher of startedWatchers) {
-          await tryStop(watcher);
-        }
-      };
-
-      // Start p2p-related services once the archiver has completed sync
-      void archiver
-        .waitForInitialSync()
-        .then(async () => {
-          for (const watcher of watchersToStart) {
-            await watcher.start();
-            startedWatchers.push(watcher);
-          }
-          log.info(`All p2p services started`);
-        })
-        .catch(err => log.error('Failed to start p2p services after archiver sync', err));
-      started.push({ stop: stopStartedWatchers });
-
-      // Validator enabled, create/start relevant service
-      let sequencer: SequencerClient | undefined;
-      let automineSequencer: AutomineSequencer | undefined;
-      let slasherClient: SlasherClientInterface | undefined;
-
-      // The slasher can run standalone to collect offenses for non-validators; it only writes to L1 when a
-      // proposer is elected (which requires a sequencer), so running it read-only on a non-validator is safe.
-      if (collectOffenses) {
-        const validatorAddresses = keyStoreManager
-          ? NodeKeystoreAdapter.fromKeyStoreManager(keyStoreManager).getAddresses()
-          : [];
-
-        slasherClient = await createSlasher(
-          config,
-          pickL1ContractAddresses(config),
-          getPublicClient(config),
-          watchers,
-          dateProvider,
-          epochCache,
-          validatorAddresses,
-          undefined, // logger
-        );
-        await slasherClient.start();
-        started.push(slasherClient);
-      }
-
-      if (!config.disableValidator && validatorClient) {
-        const l1TxUtils = config.sequencerPublisherForwarderAddress
-          ? await createForwarderL1TxUtilsFromSigners(
-              publicClient,
-              keyStoreManager!.createAllValidatorPublisherSigners(),
-              config.sequencerPublisherForwarderAddress,
-              { ...config, scope: 'sequencer' },
-              { telemetry, logger: log.createChild('l1-tx-utils'), dateProvider, kzg: Blob.getViemKzgInstance() },
-            )
-          : await createL1TxUtilsFromSigners(
-              publicClient,
-              keyStoreManager!.createAllValidatorPublisherSigners(),
-              { ...config, scope: 'sequencer' },
-              { telemetry, logger: log.createChild('l1-tx-utils'), dateProvider, kzg: Blob.getViemKzgInstance() },
-            );
-
-        // Create a funder L1TxUtils from the keystore funding account (if configured)
-        const fundingSigner = keyStoreManager?.createFundingSigner();
-        let funderL1TxUtils: L1TxUtils | undefined;
-        if (fundingSigner) {
-          const [funder] = await createL1TxUtilsFromSigners(
-            publicClient,
-            [fundingSigner],
-            { ...config, scope: 'sequencer' },
-            { telemetry, logger: log.createChild('l1-tx-utils:funder'), dateProvider },
-          );
-          funderL1TxUtils = funder;
-        }
-
-        // Create and start the sequencer client
-        const checkpointsBuilder = new CheckpointsBuilder(
-          { ...config, l1GenesisTime, slotDuration: Number(slotDuration), rollupManaLimit },
-          worldStateSynchronizer,
-          archiver,
-          dateProvider,
-          telemetry,
-          debugLogStore,
-        );
-
-        if (config.useAutomineSequencer) {
-          // Test-only path: deterministic, queue-driven sequencer for non-block-building e2e tests.
-          // See `AUTOMINE_E2E_OPTS` in `end-to-end/src/fixtures/fixtures.ts`.
-          automineSequencer = await createAutomineSequencer({
-            config,
-            l1TxUtils,
-            funderL1TxUtils,
-            publicClient,
-            rollupContract,
-            epochCache,
-            blobClient,
-            telemetry,
-            dateProvider,
-            keyStoreManager: keyStoreManager!,
-            validatorClient,
-            checkpointsBuilder,
-            globalVariableBuilder,
-            worldStateSynchronizer,
-            archiver,
-            p2pClient,
-            l1Constants: {
-              l1GenesisTime,
-              slotDuration: Number(slotDuration),
-              ethereumSlotDuration: config.ethereumSlotDuration,
-              rollupManaLimit,
-            },
-            autoSettle: config.automineEnableProveEpoch,
-            log,
-          });
-        } else {
-          sequencer = await SequencerClient.new(config, {
-            ...deps,
-            epochCache,
-            l1TxUtils,
-            funderL1TxUtils,
-            validatorClient,
-            p2pClient,
-            worldStateSynchronizer,
-            slasherClient,
-            checkpointsBuilder,
-            l2BlockSource: archiver,
-            l1ToL2MessageSource: archiver,
-            telemetry,
-            dateProvider,
-            blobClient,
-            nodeKeyStore: keyStoreManager!,
-            globalVariableBuilder,
-          });
-        }
-      }
-
-      if (!options.dontStartSequencer && sequencer) {
-        await sequencer.start();
-        started.push(sequencer);
-        log.verbose(`Sequencer started`);
-      } else if (sequencer) {
-        log.warn(`Sequencer created but not started`);
-      }
-
-      if (!options.dontStartSequencer && automineSequencer) {
-        await automineSequencer.start();
-        started.push({ stop: () => automineSequencer!.stop() });
-        log.verbose(`AutomineSequencer started`);
-      } else if (automineSequencer) {
-        log.warn(`AutomineSequencer created but not started`);
-      }
-
-      // Create prover node subsystem if enabled
-      let proverNode: ProverNode | undefined;
-      if (config.enableProverNode) {
-        proverNode = await createProverNode(config, {
-          ...deps.proverNodeDeps,
-          telemetry,
-          dateProvider,
-          archiver,
-          worldStateSynchronizer,
-          p2pClient,
-          epochCache,
-          blobClient,
-          keyStoreManager,
-        });
-
-        if (!options.dontStartProverNode) {
-          await proverNode.start();
-          started.push(proverNode);
-          log.info(`Prover node subsystem started`);
-        } else {
-          log.info(`Prover node subsystem created but not started`);
-        }
-      }
-
-      const node = new AztecNodeService(
-        config,
-        p2pClient,
-        archiver,
-        archiver,
-        archiver,
-        archiver,
-        worldStateSynchronizer,
-        sequencer,
-        proverNode,
-        slasherClient,
-        validatorsSentinel,
-        stopStartedWatchers,
-        ethereumChain.chainInfo.id,
-        config.rollupVersion,
-        globalVariableBuilder,
-        rollupContract,
-        feeProvider,
-        epochCache,
-        packageVersion,
-        peerProofVerifier,
-        rpcProofVerifier,
-        telemetry,
-        log,
-        blobClient,
-        validatorClient,
-        keyStoreManager,
-        debugLogStore,
-        automineSequencer,
-      );
-
-      return node;
-    } catch (err) {
-      log.error('Failed during node creation, stopping started resources', err);
-      for (const resource of started.reverse()) {
-        await tryStop(resource);
-      }
-      throw err;
-    }
-  }
-
-  /**
-   * Verifies the node's configured L1 timing matches the rollup contract it is pointed at, for the fields the
-   * node's own config carries. Each comparison is guarded against an undefined config value, so a config that
-   * does not carry a field is not checked. Throws a single error listing every mismatch. Runs in the shared
-   * startup path for every node role.
-   */
-  private static checkConfigMatchesRollup(
-    config: AztecNodeConfig,
-    rollup: { slotDuration: number; epochDuration: number },
-  ): void {
-    const mismatches: string[] = [];
-    if (config.aztecSlotDuration !== undefined && config.aztecSlotDuration !== rollup.slotDuration) {
-      mismatches.push(`aztecSlotDuration is ${config.aztecSlotDuration} but the rollup reports ${rollup.slotDuration}`);
-    }
-    if (config.aztecEpochDuration !== undefined && config.aztecEpochDuration !== rollup.epochDuration) {
-      mismatches.push(
-        `aztecEpochDuration is ${config.aztecEpochDuration} but the rollup reports ${rollup.epochDuration}`,
-      );
-    }
-    if (mismatches.length > 0) {
-      throw new Error(
-        `The node's configured L1 timing does not match the rollup contract it is pointed at: ${mismatches.join('; ')}`,
-      );
-    }
   }
 
   /**
