+++ yarn-project/aztec-node/src/index.ts
@@ -1,3 +1,4 @@
 export * from './aztec-node/config.js';
 export * from './aztec-node/register_node_rpc_handlers.js';
 export * from './aztec-node/server.js';
+export * from './factory.js';
